pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crossfire"
version = "0.1.0"
description = "Regex rule overlap analyzer for DLP, secret scanning, SAST, and IDS tools"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Crossfire Contributors"}
]
keywords = ["regex", "overlap", "dedup", "dlp", "secret-scanning", "sast", "rules"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "rstr>=3.2",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
rich = ["rich>=13.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
    "types-PyYAML",
    "pre-commit",
]

[project.scripts]
crossfire = "crossfire.cli:main"

[project.entry-points."crossfire.adapters"]

[project.urls]
Homepage = "https://github.com/lumen-argus/crossfire"
Repository = "https://github.com/lumen-argus/crossfire"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "C4",     # flake8-comprehensions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "RUF",    # ruff-specific rules
    "PERF",   # perflint
]
ignore = [
    "PERF203",  # try-except in loop — intentional for per-item error handling
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PT011"]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["crossfire"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["tomllib", "tomli", "sre_parse", "re._parser"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v", "--strict-markers"]
